+++ src/gis/fit/defs/CFitProfile.h
@@ -32,11 +32,11 @@ public:
     CFitProfile(QString name, quint16 globalMesgNr);
     virtual ~CFitProfile();
 
-    void addField(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, quint16 offset, QString units);
+    void addField(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, qint16 offset, QString units);
     void addSubfield(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal
                      scale, quint16 offset, QString units, quint8 subRefFieldDefNr, quint8 subRefFieldValue);
-    void addComponent(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, quint16 offset, QString units, quint8 componentFieldDefNr, quint8 bits);
-    void addComponent(int subfieldIndex, QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, quint16 offset, QString units, quint8 componentFieldDefNr, quint8 bits);
+    void addComponent(QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, qint16 offset, QString units, quint8 componentFieldDefNr, quint32 bits);
+    void addComponent(int subfieldIndex, QString name, const CFitBaseType& baseType, quint8 fieldDefNr, qreal scale, qint16 offset, QString units, quint8 componentFieldDefNr, quint32 bits);
 
 
     const CFitFieldProfile* getField(quint8 fieldDefNr) const;
